backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
PyJWT==2.8.0
werkzeug==3.0.1
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.0
joblib==1.4.2

